+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [WIP] 0.3.1 / 2017-June-??
 - [TODO] Use Bract 0.3.1 components
-- [TODO] Rename repo to `demo.wordcount`
+- Rename project to `demo.wordcount`
 
 
 ## 0.3.0 / 2017-June-11