+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# wordcount TODO and Change Log
+# demo.wordcount Change Log
+
+
+## [WIP] 0.3.1 / 2017-June-??
+- Use Bract 0.3.1 components
+- Rename project to `demo.wordcount`
 
 
 ## 0.3.0 / 2017-June-11